Modal dismiss focuses incorrect button
The bootstrap accessibitlity gem attempts to set the focus of the parent
button that is fired the modal. This functionality is fixed in bootstrap
and the accessibility gem is causing issues when there are multiple
buttons that fire the same modal. The code focuses the last rendered
button on the page.

See example http://getbootstrap.com/javascript/#modals-related-target

Tested With:
Internet Explorer 11
Firefox ESR 38.4.0
Chrome 46.0.2490.86 (64-bit)
Safari 9.0.1

Bootstrap Accessibility v1.0.4
Bootstrap  3.3.2